Split the register flow's organization specs into their own file
The two organization cases - registering with the one you belong to, and
an organization's e-vehicle safety rules - were ~a third of the file's
wall clock. In their own file they can run alongside the rest.
